examples/tgr2023_06_imglabel/requirements.txt
pyserial
pyside6
opencv-contrib-python
numpy